+++ source/ProcessManager.Example.Orchestrations.Tests/Integration/Processes/BRS_X01/NoInputExample/V1/MonitorOrchestrationUsingClientScenario.cs
@@ -102,8 +102,11 @@ public class MonitorOrchestrationUsingClientScenario : IAsyncLifetime
                         CancellationToken.None);
 
                 return
-                    orchestrationInstance.Lifecycle.State == OrchestrationInstanceLifecycleState.Terminated
-                    && orchestrationInstance.Lifecycle.TerminationState == OrchestrationInstanceTerminationState.Succeeded;
+                    orchestrationInstance.Lifecycle is
+                    {
+                        State: OrchestrationInstanceLifecycleState.Terminated,
+                        TerminationState: OrchestrationInstanceTerminationState.Succeeded
+                    };
             },
             timeLimit: TimeSpan.FromSeconds(60),
             delay: TimeSpan.FromSeconds(3));
@@ -124,5 +127,10 @@ public class MonitorOrchestrationUsingClientScenario : IAsyncLifetime
                 CancellationToken.None);
 
         orchestrationInstancesGeneralSearch.Should().Contain(x => x.Id == orchestrationInstanceId);
+        orchestrationInstancesGeneralSearch.Should().ContainSingle();
+        var orchestrationInstance = orchestrationInstancesGeneralSearch.Single();
+        orchestrationInstance.ActorMessageId.Should().BeNull();
+        orchestrationInstance.TransactionId.Should().BeNull();
+        orchestrationInstance.MeteringPointId.Should().BeNull();
     }
 }