+++ source/ProcessManager.Example.Orchestrations/Processes/BRS_X01/InputExample/V1/StartInputExampleHandlerV1.cs
@@ -17,6 +17,7 @@ using Energinet.DataHub.ProcessManager.Core.Application.Orchestration;
 using Energinet.DataHub.ProcessManager.Core.Domain.OrchestrationDescription;
 using Energinet.DataHub.ProcessManager.Core.Domain.OrchestrationInstance;
 using Energinet.DataHub.ProcessManager.Example.Orchestrations.Abstractions.Processes.BRS_X01.InputExample.V1.Model;
+using Energinet.DataHub.ProcessManager.Example.Orchestrations.Processes.BRS_X01.InputExample.V1.Steps;
 using NodaTime.Extensions;
 
 namespace Energinet.DataHub.ProcessManager.Example.Orchestrations.Processes.BRS_X01.InputExample.V1;
@@ -32,7 +33,7 @@ internal class StartInputExampleHandlerV1(
     {
         // Here we show how its possible, based on input, to decide certain steps should be skipped by the orchestration.
         IReadOnlyCollection<int> skipStepsBySequence = command.InputParameter.ShouldSkipSkippableStep
-            ? [Orchestration_Brs_X01_InputExample_V1.SkippableStepSequence]
+            ? [SkippableStep.StepSequence]
             : [];
 
         var orchestrationInstanceId = await _manager
@@ -54,7 +55,7 @@ internal class StartInputExampleHandlerV1(
     {
         // Here we show how its possible, based on input, to decide certain steps should be skipped by the orchestration.
         IReadOnlyCollection<int> skipStepsBySequence = command.InputParameter.ShouldSkipSkippableStep
-            ? [Orchestration_Brs_X01_InputExample_V1.SkippableStepSequence]
+            ? [SkippableStep.StepSequence]
             : [];
 
         var orchestrationInstanceId = await _manager