+++ source/ProcessManager.Orchestrations/Processes/BRS_021/ForwardMeteredData/V1/Mapper/MeteringPointMasterDataMapper.cs
@@ -12,47 +12,47 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using EMTypes = Energinet.DataHub.ElectricityMarket.Integration;
+using Energinet.DataHub.ElectricityMarket.Integration.Models.MasterData;
 using PMTypes = Energinet.DataHub.ProcessManager.Components.Abstractions.ValueObjects;
 
 namespace Energinet.DataHub.ProcessManager.Orchestrations.Processes.BRS_021.ForwardMeteredData.V1.Mapper;
 
 public static class MeteringPointMasterDataMapper
 {
-    public static readonly Dictionary<EMTypes.MeteringPointType, PMTypes.MeteringPointType> MeteringPointTypeMap = new()
+    public static readonly Dictionary<MeteringPointType, PMTypes.MeteringPointType> MeteringPointTypeMap = new()
     {
-        { EMTypes.MeteringPointType.Consumption, PMTypes.MeteringPointType.Consumption },
-        { EMTypes.MeteringPointType.Production, PMTypes.MeteringPointType.Production },
-        { EMTypes.MeteringPointType.Exchange, PMTypes.MeteringPointType.Exchange },
+        { MeteringPointType.Consumption, PMTypes.MeteringPointType.Consumption },
+        { MeteringPointType.Production, PMTypes.MeteringPointType.Production },
+        { MeteringPointType.Exchange, PMTypes.MeteringPointType.Exchange },
     };
 
-    public static readonly Dictionary<EMTypes.MeasureUnit, PMTypes.MeasurementUnit> MeasureUnitMap = new()
+    public static readonly Dictionary<MeasureUnit, PMTypes.MeasurementUnit> MeasureUnitMap = new()
     {
-        { EMTypes.MeasureUnit.Ampere, PMTypes.MeasurementUnit.Ampere },
-        { EMTypes.MeasureUnit.STK, PMTypes.MeasurementUnit.Pieces },
-        { EMTypes.MeasureUnit.kVArh, PMTypes.MeasurementUnit.KiloVoltAmpereReactiveHour },
-        { EMTypes.MeasureUnit.kWh, PMTypes.MeasurementUnit.KilowattHour },
-        { EMTypes.MeasureUnit.kW, PMTypes.MeasurementUnit.Kilowatt },
-        { EMTypes.MeasureUnit.MW, PMTypes.MeasurementUnit.Megawatt },
-        { EMTypes.MeasureUnit.MWh, PMTypes.MeasurementUnit.MegawattHour },
-        { EMTypes.MeasureUnit.Tonne, PMTypes.MeasurementUnit.MetricTon },
-        { EMTypes.MeasureUnit.MVAr, PMTypes.MeasurementUnit.MegaVoltAmpereReactivePower },
-        { EMTypes.MeasureUnit.DanishTariffCode, PMTypes.MeasurementUnit.DanishTariffCode },
+        { MeasureUnit.Ampere, PMTypes.MeasurementUnit.Ampere },
+        { MeasureUnit.STK, PMTypes.MeasurementUnit.Pieces },
+        { MeasureUnit.kVArh, PMTypes.MeasurementUnit.KiloVoltAmpereReactiveHour },
+        { MeasureUnit.kWh, PMTypes.MeasurementUnit.KilowattHour },
+        { MeasureUnit.kW, PMTypes.MeasurementUnit.Kilowatt },
+        { MeasureUnit.MW, PMTypes.MeasurementUnit.Megawatt },
+        { MeasureUnit.MWh, PMTypes.MeasurementUnit.MegawattHour },
+        { MeasureUnit.Tonne, PMTypes.MeasurementUnit.MetricTon },
+        { MeasureUnit.MVAr, PMTypes.MeasurementUnit.MegaVoltAmpereReactivePower },
+        { MeasureUnit.DanishTariffCode, PMTypes.MeasurementUnit.DanishTariffCode },
     };
 
-    public static readonly Dictionary<EMTypes.MeteringPointSubType, Model.MeteringPointSubType> MeteringPointSubTypeMap = new()
+    public static readonly Dictionary<MeteringPointSubType, Model.MeteringPointSubType> MeteringPointSubTypeMap = new()
     {
-        { EMTypes.MeteringPointSubType.Physical, Model.MeteringPointSubType.Physical },
-        { EMTypes.MeteringPointSubType.Virtual, Model.MeteringPointSubType.Virtual },
-        { EMTypes.MeteringPointSubType.Calculated, Model.MeteringPointSubType.Calculated },
+        { MeteringPointSubType.Physical, Model.MeteringPointSubType.Physical },
+        { MeteringPointSubType.Virtual, Model.MeteringPointSubType.Virtual },
+        { MeteringPointSubType.Calculated, Model.MeteringPointSubType.Calculated },
     };
 
-    public static readonly Dictionary<EMTypes.ConnectionState, Model.ConnectionState> ConnectionStateMap = new()
+    public static readonly Dictionary<ConnectionState, Model.ConnectionState> ConnectionStateMap = new()
     {
-        { EMTypes.ConnectionState.NotUsed, Model.ConnectionState.NotUsed },
-        { EMTypes.ConnectionState.ClosedDown, Model.ConnectionState.ClosedDown },
-        { EMTypes.ConnectionState.New, Model.ConnectionState.New },
-        { EMTypes.ConnectionState.Connected, Model.ConnectionState.Connected },
-        { EMTypes.ConnectionState.Disconnected, Model.ConnectionState.Disconnected },
+        { ConnectionState.NotUsed, Model.ConnectionState.NotUsed },
+        { ConnectionState.ClosedDown, Model.ConnectionState.ClosedDown },
+        { ConnectionState.New, Model.ConnectionState.New },
+        { ConnectionState.Connected, Model.ConnectionState.Connected },
+        { ConnectionState.Disconnected, Model.ConnectionState.Disconnected },
     };
 }